requirements.txt
Django==1.7.1
argparse==1.2.1
distribute==0.6.24
djangorestframework==3.0.2
simplejson==3.6.5
wsgiref==0.1.2
